requirements.txt
anthropic>=0.40.0
flask>=3.0.3
chromadb>=0.5.5
sentence-transformers>=3.1.1
python-dotenv>=1.0.1
gunicorn>=22.0.0
supabase>=2.0.0
